[Ide] Fixup rebuild progress monitor behaviour
We need to have one statusbar progress per action we're doing, so we
will have to attach the same error pad to a different status bar for
build and clean.

The idea of the code is to have one statusbar monitor alive at a time,
while the error one is alive for the whole operation, and the
aggregated monitors are disposed at the end just so they're not
marked as finalized but not Disposed by profiler tools.